# p2/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Books API",
    "description": "API for managing books - CSE 341",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "sessionAuth": {
      "type": "apiKey",
      "in": "cookie",
      "name": "connect.sid",
      "description": "Session cookie set after logging in via GET /auth/github. Log in through a browser first, then use Swagger UI in that SAME browser tab so the cookie is sent automatically with \"Try it out\"."
    }
  },
  "paths": {
...
    "/auth/profile": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current logged-in user",
        "description": "Returns the profile of the currently authenticated user. Requires an active session — log in via GET /auth/github in a browser first, then call this from the same browser session.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ]
      }
    },
...
  }
}